Day 8 part 2

+++ src/lib.rs
@@ -40,7 +40,7 @@ pub fn solve(day: u8, part: Part, input: &str) -> u32 {
         (4, Part::Two) => day_4::solve_part2(input),
         (1, Part::Two) => day_1::solve_part2(input),
         (8, Part::One) => day_8::solve_part1(input),
-        (8, Part::Two) => day_8_part2::solve(input),
+        (8, Part::Two) => day_8_part2::solve(input) as u32,
         _ => todo!(),
     }
 }
